requirements.txt
Flask==3.0.3
Pillow==10.4.0
piexif==1.1.3
numpy==1.26.4
scikit-image==0.24.0

